StEP00086: Sperrung von Veranstaltung gegen Selbsteintrag

+++ public/admin_admission.php
@@ -135,7 +135,7 @@ if ($SessSemName[1])
 	$seminar_id=$SessSemName[1];
 
 //wenn wir frisch reinkommen, werden benoetigte Daten eingelesen
-if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm_chrono_x) && (!$add_studg_x) && (!$delete_studg)) {
+if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm_chrono_x) && (!$add_studg_x) && (!$delete_studg) && (!$adm_gesperrt_x)) {
 	$db->query("SELECT * FROM seminare WHERE Seminar_id = '$seminar_id' ");
 	$db->next_record();
 	$admin_admission_data='';
@@ -230,6 +230,9 @@ if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm
 	if ($adm_chrono_x)
 		$admin_admission_data["admission_type"]=2;
 
+	if ($adm_gesperrt_x)
+                $admin_admission_data["admission_type"] = 3;
+
 	//Aenderungen ubernehmen
 	$admin_admission_data["admission_binding"]=$admission_binding;
 	if ($admin_admission_data["admission_binding"])
@@ -323,9 +326,9 @@ if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm
 		}
 
 	//Checks bei Anmeldeverfahren
-	} elseif ((!$adm_chrono_x) && (!$adm_los_x))  {
+	} elseif ((!$adm_chrono_x) && (!$adm_los_x) && (!$adm_gesperrt_x))  {
 		//max. Teilnehmerzahl checken
-		if (($uebernehmen_x) && ($admin_admission_data["admission_type"])) {
+		if (($uebernehmen_x) && ($admin_admission_data["admission_type"] > 0) && ($admin_admission_data["admission_type"] < 3)) {
 			if ($admin_admission_data["admission_turnout"] < 1) {
 				$errormsg=$errormsg."error§"._("Wenn Sie die Teilnahmebeschr&auml;nkung benutzen wollen, m&uuml;ssen Sie wenigstens einen Teilnehmer zulassen.")."§";
 				$admin_admission_data["admission_turnout"] =1;
@@ -367,7 +370,7 @@ if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm
 
 		//Ende der Anmeldung checken
 		if ($uebernehmen_x)
-			if (($admin_admission_data["admission_type"]) && ($admin_admission_data["admission_endtime"])) {
+			if (($admin_admission_data["admission_type"]) && ($admin_admission_data["admission_endtime"]) && ($admin_admission_data["admission_type"]!=3)) {
 				if ($admin_admission_data["admission_type"] == 1)
 					$end_date_name="Losdatum";
 				else
@@ -498,7 +501,7 @@ if (($seminar_id) && (!$uebernehmen_x) &&(!$adm_null_x) &&(!$adm_los_x) &&(!$adm
 			}
 
 		//Variante nachtraeglich Anmeldeverfahren starten, alle alten Teilnehmer muessen raus
-		if (($admin_admission_data["admission_type"] >$admin_admission_data["admission_type_org"]) && ($admin_admission_data["admission_type_org"]==0)) {
+		if (($admin_admission_data["admission_type"] >$admin_admission_data["admission_type_org"]) && ($admin_admission_data["admission_type_org"]==0) && $admin_admission_data["admission_type"]!=3) {
 			$db->query("SELECT seminar_user.user_id, username FROM seminar_user LEFT JOIN auth_user_md5 USING (user_id) WHERE Seminar_id ='".$admin_admission_data["sem_id"]."' AND status IN ('autor', 'user') ");
 			$db2->query("DELETE FROM seminar_user WHERE Seminar_id ='".$admin_admission_data["sem_id"]."' AND status IN ('autor', 'user') ");
 			if ($db2->affected_rows()) {
@@ -647,10 +650,11 @@ if (is_array($admin_admission_data["studg"]) && $admin_admission_data["admission
 						printf ("<li><font size=-1><a href=\"about.php?username=%s\">%s</a></font></li>", $db->f("username"), htmlReady($db->f("fullname")));
 					}
 				} else { ?>
-				<font size=-1><?=_("Sie k&ouml;nnen hier eine Teilnahmebeschr&auml;nkung per Anmeldeverfahren festlegen. Sie k&ouml;nnen per Losverfahren beschr&auml;nken oder Anmeldungen in der Reihenfolge ihres Eintreffens (chronologische Anmeldung) zulassen.")?><br /></font>
+				<font size=-1><?=_("Sie k&ouml;nnen hier eine Teilnahmebeschr&auml;nkung per Anmeldeverfahren festlegen. Sie k&ouml;nnen per Losverfahren beschr&auml;nken oder Anmeldungen in der Reihenfolge ihres Eintreffens (chronologische Anmeldung) zulassen. Wenn Sie eine Veranstaltung sperren, kann sich niemand zu dieser Veranstaltung anmelden. Bestehende Teilnehmer- und Wartelisteneintr&auml;ge bleiben bei einem Wechsel von <B>keins</B> auf <B>gesperrt</B> unber&uuml;hrt.")?><br /></font>
 				<br /><input type="IMAGE" name="adm_null" <? if  ($admin_admission_data["admission_type"] == 0) print makeButton ("keins2", "src");  else print makeButton ("keins", "src") ?> border=0 value="keins">&nbsp;
 				<input type="IMAGE" name="adm_los" <? if  ($admin_admission_data["admission_type"] == 1) print makeButton ("los2", "src");  else print makeButton ("los", "src") ?> border=0 value="los">&nbsp;
 				<input type="IMAGE" name="adm_chrono" <? if  ($admin_admission_data["admission_type"] == 2) print makeButton ("chronolog2", "src");  else print makeButton ("chronolog", "src") ?>border=0 value="chronolog">
+				<input type="IMAGE" name="adm_gesperrt" <? if  ($admin_admission_data["admission_type"] == 3) print makeButton ("gesperrt2", "src");  else print makeButton ("gesperrt", "src") ?>border=0 value="gesperrt">
 				<? } ?>
 				<input type="HIDDEN" name="adm_type_old" value="<? echo $admin_admission_data["admission_type"] ?>" /><br />
 			</td>
@@ -733,7 +737,7 @@ if (is_array($admin_admission_data["studg"]) && $admin_admission_data["admission
 			</tr>
 		<?
 		}
-		if (!$admin_admission_data["admission_type"]) {
+		if (!$admin_admission_data["admission_type"]  || $admin_admission_data["admission_type"] == 3) {
 		?>
 		<tr <? $cssSw->switchClass() ?>>
 			<td class="<? echo $cssSw->getClass() ?>" width="4%" align="right">
